Create a new release on every pull request merge
Using semantic versioning[1]

Use GoReleaser[2] to create a new release every time a pull request is
merged to master.
This requires each pull request to have a semantic version label, which
we enforce by using a label-checker GitHub Action[3].

[1] https://semver.org/
[2] https://goreleaser.com/ci/
[3] https://github.com/agilepathway/label-checker

NB we don't build an executable when releasing, as this is not
(currently, anyway) a Go project.