Add end-to-end tests
The tests run the actual GitHub Action and then verify that Hoverfly and
Hoverctl have been successfully installed, or not.

This provides great end-to-end black box testing[1] coverage :-)

GitHub Actions cannot be run locally[2], so that means that these tests
cannot be run locally either.  Instead, they run automatically as a
GitHub Action themselves[3], triggered on every push.

There is no need for a separate language for the tests - as we
are running the actual GitHub Action we are able use the GitHub Action
workflow syntax[4], which gives us what we need (e.g. expressions)[5] to
write clean tests.

[1] http://softwaretestingfundamentals.com/black-box-testing
[2] https://github.community/t/can-i-run-github-actions-on-my-laptop/17019/2
[3] https://github.com/agilepathway/hoverfly-github-action/actions?query=workflow%3ATest
[4] https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions
[5] https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions#about-contexts-and-expressions