Ensure dependabot notifies of new versions
Dependabot does not currently update [references to Docker
Containers][1] in GitHub Actions.  So this is a hack to ensure that
Dependabot still notifies us (via Pull Requests) whenever the [Label
Checker][2] is updated.

We never want this fake GitHub Action to be triggered (because the
Dependabot doesn't need the action to be triggered in order to carry out
it's version updating), so we have set it to only ever trigger on a
branch name which we will never have.

The PR that Dependabot creates will only have the update for this file,
so we'll have to then manually update the version in the
`check_semver_labels.yml` too.

[1]: https://dependabot.com/github-actions/
[2]: https://github.com/agilepathway/label-checker