Use Dependabot for Hoverfly version notifications (#30)
Dependabot will create a PR whenever there is a new version of Hoverfly
released. We will then amend the PR by manually updating to the new
version in the various places in the code where it is used. We could in
theory automate this by, for instance, having a GitHub Action that does
it, triggered by the Dependabot PR, but let's start simple - it could
well be overkill to do this.  Especially as Dependabot itself may
provide a solution to this one day.