Use a pinned devcontainer base image version
We want to pin to the most specific version we can, to ensure that the
devcontainer builds are always reproducible.

This is despite the general guidance from Microsoft for these images
being to use ["the major release version of this tag to avoid breaking
changes while still taking fixes and content additions as they land.
e.g. `0-buster`"][1]

We will still take fixes and content additions as they land, as we use
Dependabot to keep us up to date.  Dependabot will create a PR whenever
there is a new version of the base image, and we then manually apply
that new version to the `devcontainer.json` (full details on how to do
this are in the `DEPENDENCIES.md`).

See also [this documentation on the base image versions][2]

[1]: https://hub.docker.com/_/microsoft-vscode-devcontainers?tab=description
[2]: https://github.com/microsoft/vscode-dev-containers/tree/master/containers/debian#using-this-definition-with-an-existing-folder