Fix bug where tests were not running
For some reason the tests file (tests.yml) was not being picked up by
GitHub Actions, meaning that the tests were not being run.  Renaming the
file to `test.yml` does trigger the tests, so this commit renames the
file.  Don't know why `tests.yml` is not picked up, maybe it's now a
reserved filename by GitHub or something.