Fix add path bug (#66)
* Fix bug where tests were not running

For some reason the tests file (tests.yml) was not being picked up by
GitHub Actions, meaning that the tests were not being run.  Renaming the
file to `test.yml` does trigger the tests, so this commit renames the
file.  Don't know why `tests.yml` is not picked up, maybe it's now a
reserved filename by GitHub or something.

* Fix add path bug

GitHub [changed the functionality for adding a path][1], in response
to a [security vulnerability][2].  This was a breaking change, which
broke our Hoverfly GitHub Action.  This commit fixes the breakage by
using [the new way of adding a path][3].

[1]: https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/
[2]: https://github.com/actions/toolkit/security/advisories/GHSA-mfwh-5m23-j46w
[3]: https://docs.github.com/en/actions/reference/workflow-commands-for-github-actions#adding-a-system-path