Use ubuntu-latest in GitHub Actions (#73)
Prior to this commit we had harcoded the Ubuntu version in our GitHub
Actions to 20.04, with a GitHub Issue being automatically generated
every year to update to the latest version.

GitHub Actions only updates every 2 years, it turns out, not ever year,
as they only upgrade to LTS versions.  GitHub Actions don't support a
Cron scheduler in the format that's needed for every 2 years, so have
decided to be pragmatic and just use ubuntu-latest.