Use latest dependency versions in devcontainer
Prior to this commit we were pinning specific apt-get versions in the
devcontainer, in order to avoid the devcontainer breaking due to a
version of a dependency changing.  In practice the risk of this is very
small though, and the cost of using pinned apt-get versions was
outweighing the perceived benefit.