Use the Go minor version for Dependabot updates (#108)
Prior to this commit Dependabot would create a pull request for every
new patch version of Go. In `github_tag_and_release.yml` we only specify
the minor version of Go though, so it is better to have Dependabot only
create pull requests when there is a new minor version.